backend/requirements.txt
# Web API
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
pymongo==4.10.1

# Core LangChain unified versions
langchain==0.3.12
langchain-community==0.3.12
langchain-openai==0.2.3

# Vector DB
faiss-cpu==1.8.0

# Azure/OpenAI + tooling
openai==1.55.3
langsmith==0.1.96

# Required dependencies pinned for Python 3.13
numpy==2.1.3
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
pymongo>=4.10.1


